+++ lib/jets/overrides/rails.rb
@@ -3,3 +3,4 @@ require "jets/overrides/rails/url_helper"
 require "jets/overrides/rails/rendering_helper"
 require "jets/overrides/rails/asset_tag_helper"
 require "jets/overrides/rails/action_controller"
+require "jets/overrides/rails/migration_checker"
