+++ lib/jets/spec_helpers.rb
@@ -10,9 +10,8 @@ module Jets
 end
 
 unless ENV["SKIP_MIGRATION_CHECK"] == "true"
-  require_relative './core_ext/overrides/active_record/migration_checker'
   ActiveRecord::Tasks::DatabaseTasks.db_dir = "#{Jets.root}/db"
-  ActiveRecord::Migration.extend Overrides::ActiveRecord::MigrationChecker
+  ActiveRecord::Migration.extend ActiveRecord::MigrationChecker
   ActiveRecord::Migration.prepare_test_db
 end
